Move functions coding for uuid to separate file
